# openapi.yaml
openapi: 3.1.0
info:
  title: Ticket API
  version: 0.1.891
  description: |
    RESTful API for the Ticket project and issue management system.
    All endpoints are under `/api/`. Authentication uses Bearer tokens
    or session cookies (`__Host-session` on secure requests; legacy
    `ticket_token` remains accepted).
    Agent endpoints use HTTP Basic Auth (agent UUID + password).
  contact:
    name: Simon Gauld
    url: https://github.com/simonski/ticket
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from /api/login
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Host-session
    BasicAuth:
      type: http
      scheme: basic
      description: Agent UUID as username, agent password as password
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
      example:
        error: "ticket not found"
...
    CountSummary:
      type: object
      description: Aggregate counts by type and status
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/config/registration:
    post:
      tags: [Config]
      summary: Enable or disable user registration
      operationId: setRegistration
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Registration setting updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  registration_enabled:
                    type: boolean
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/config/chat_enabled:
    post:
      tags: [Config]
      summary: Enable or disable chat feature
      operationId: setChatEnabled
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Chat setting updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  chat_enabled:
                    type: boolean

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/config/chat_limits:
    post:
      tags: [Config]
      summary: Set chat connection and duration limits
      operationId: setChatLimits
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                max_connections:
                  type: integer
                  default: 10
                max_duration_minutes:
                  type: integer
                  default: 30
      responses:
        '200':
          description: Chat limits updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  chat_max_connections:
                    type: integer
                  chat_max_duration_minutes:
                    type: integer

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  # ── Users ───────────────────────────────────────────────────────────
...
  /api/count:
    get:
      tags: [Tickets]
      summary: Get aggregate counts by type and status
      operationId: getCountSummary
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
            format: int64
          description: Filter to a specific project
      responses:
        '200':
          description: Count summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountSummary'

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...